services/explorers/src/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
beanie==1.23.6
gpxpy==1.6.1
click==8.1.7
dnspython==2.4.2
fastapi==0.104.1
h11==0.14.0
idna==3.4
lazy-model==0.2.0
motor==3.3.1
pydantic==2.5.0
pydantic_core==2.14.1
pydantic-settings==2.1.0
PyJWT==2.8.0
pymongo==4.6.0
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
python-multipart==0.0.6
typing_extensions==4.8.0
uvicorn==0.24.0.post1
lxml==4.9.3